Added serial evisceration logic
As of right now, this code assumes that multiple eviscerators per zone means that the hosts to be eviscerated are to be 'divided equally' across those eviscerators in parallel, and not multiply eviscerated via those multiple machines. Furthermore, this codebase excludes hosts from being eviscerated once they have been 'eviscerated'.

A roundabout to this is to separate eviscerators (assuming there is no cross contamination) into their zones (1 zone = 1 eviscerator machine) and that the zones feed into one another. AND, you need to set SERIAL _EVISCERATION to be true, such that upon transport, any eviscerated chickens become considered, NOT 'eviscerated', so they can be 'eviscerated' again. 

Some may desire this feature since evisceration often can happen withmultiple probing machines (1 to create the hole, 1 to check and resize the hole, and 1 finally to eviscerate etc).